backend/requirements.txt
# ── Web Framework ───────────────────────────
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ── Database ─────────────────────────────────
# 2.0.31+ — исправление TypingOnly / Python 3.13 (AssertionError при импорте)
sqlalchemy[asyncio]==2.0.44
asyncpg==0.30.0
alembic==1.12.1
psycopg2-binary==2.9.9

# ── Redis ────────────────────────────────────
redis[hiredis]==5.0.1

# ── Task Queue ───────────────────────────────
dramatiq[rabbitmq,watch]==1.15.0
pika==1.3.2

# ── File Storage ─────────────────────────────
minio==7.2.0

# ── Auth ─────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# ── Validation ───────────────────────────────
# 2.10+ даёт готовые wheels pydantic-core под Python 3.13 (Windows без MSVC).
pydantic==2.12.4
pydantic-settings==2.10.1
email-validator==2.2.0

# ── Email ────────────────────────────────────
aiosmtplib==3.0.0
jinja2==3.1.2

# ── Utils ────────────────────────────────────
python-dotenv==1.2.2
httpx==0.25.2

# ── Monitoring ───────────────────────────────
prometheus-fastapi-instrumentator==6.1.0

# ── Dev / Test ───────────────────────────────
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
